firesidegames/requirements.txt
Django>=3.2,<3.3
django-allauth>=0.45.0
psycopg2>=2.9.1
python-dotenv>=0.18.0
google-cloud-storage>=1.42.2
django-storages>=1.11.1
django-sendgrid-v5>=1.1.1
django-environ>=0.7.0
google-cloud-secret-manager>=2.7.2
pytz>=2021.1
sentry-sdk>=1.4.3
colorlog>=5.0.1
Pillow>=8.3.2